moving symfony and doctrine rules
