+++ config/doctrine-rules.neon
@@ -1,0 +1,4 @@
+rules:
+    - Symplify\PHPStanRules\PHPStan\Rules\Doctrine\NoGetRepositoryOutsideServiceRule
+    - Symplify\PHPStanRules\PHPStan\Rules\Doctrine\NoParentRepositoryRule
+    - Symplify\PHPStanRules\PHPStan\Rules\Doctrine\NoRepositoryCallInDataFixtureRule
