+++ src/Doctrine/DoctrineEntityDocumentAnalyser.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace TomasVotruba\Handyman\PHPStan;
+namespace Symplify\PHPStanRules\PHPStan;
 
 use PHPStan\PhpDoc\ResolvedPhpDocBlock;
 use PHPStan\Reflection\ClassReflection;
