+++ src/Rules/Symfony/NoStringInGetSubscribedEventsRule.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace TomasVotruba\Handyman\PHPStan\Rule;
+namespace Symplify\PHPStanRules\PHPStan\Rule;
 
 use PhpParser\Node;
 use PhpParser\Node\Expr;
@@ -15,7 +15,7 @@ use PHPStan\Reflection\ClassReflection;
 use PHPStan\Rules\Rule;
 use PHPStan\Rules\RuleError;
 use PHPStan\Rules\RuleErrorBuilder;
-use TomasVotruba\Handyman\Enum\ClassName;
+use Symplify\PHPStanRules\Enum\ClassName;
 
 /**
  * @implements Rule<ClassMethod>
