+++ tests/Rules/Doctrine/NoGetRepositoryOutsideServiceRule/Fixture/NonRepositoryUsingEntityManager.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace TomasVotruba\Handyman\Tests\PHPStan\Rule\NoGetRepositoryOutsideServiceRule\Fixture;
+namespace Symplify\PHPStanRules\Tests\Rules\Doctrine\NoGetRepositoryOutsideServiceRule\Fixture;
 
 use Doctrine\ORM\EntityManager;
