+++ tests/Rules/Doctrine/NoRepositoryCallInDataFixtureRule/Fixture/SkipNonFixtureClass.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace TomasVotruba\Handyman\Tests\PHPStan\Rule\NoRepositoryCallInDataFixtureRule\Fixture;
+namespace Symplify\PHPStanRules\Tests\Rules\Doctrine\NoRepositoryCallInDataFixtureRule\Fixture;
 
 use Doctrine\Persistence\ObjectManager;
