+++ tests/Rules/PHPUnit/NoMockOnlyTestRule/Source/FirstClass.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace TomasVotruba\Handyman\Tests\PHPStan\Rule\NoMockOnlyTestRule\Source;
+namespace Symplify\PHPStanRules\Tests\PHPStan\Rule\NoMockOnlyTestRule\Source;
 
 final class FirstClass
 {
