+++ tests/Rules/Symfony/NoAbstractControllerConstructorRule/Fixture/SkipNonAbstractController.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace TomasVotruba\Handyman\Tests\PHPStan\Rule\NoAbstractControllerConstructorRule\Fixture;
+namespace Symplify\PHPStanRules\Tests\PHPStan\Rule\NoAbstractControllerConstructorRule\Fixture;
 
 final class SkipNonAbstractController
 {
