+++ tests/Rules/Symfony/NoRequiredOutsideClassRule/Fixture/TraitWithRequireAttribute.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace TomasVotruba\Handyman\Tests\PHPStan\Rule\NoRequiredOutsideClassRule\Fixture;
+namespace Symplify\PHPStanRules\Tests\PHPStan\Rule\NoRequiredOutsideClassRule\Fixture;
 
 use Symfony\Contracts\Service\Attribute\Required;
