+++ tests/Rules/Symfony/NoRequiredOutsideClassRule/NoRequiredOutsideClassRuleTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace TomasVotruba\Handyman\Tests\PHPStan\Rule\NoRequiredOutsideClassRule;
+namespace Symplify\PHPStanRules\Tests\PHPStan\Rule\NoRequiredOutsideClassRule;
 
 use Iterator;
 use PHPStan\Rules\Rule;
 use PHPStan\Testing\RuleTestCase;
 use PHPUnit\Framework\Attributes\DataProvider;
-use TomasVotruba\Handyman\PHPStan\Rule\NoRequiredOutsideClassRule;
+use Symplify\PHPStanRules\PHPStan\Rule\NoRequiredOutsideClassRule;
 
 final class NoRequiredOutsideClassRuleTest extends RuleTestCase
 {
