+++ tests/Rules/Symfony/SingleArgEventDispatchRule/Fixture/SkipUnrelatedDispatch.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace TomasVotruba\Handyman\Tests\PHPStan\Rule\SingleArgEventDispatchRule\Fixture;
+namespace Symplify\PHPStanRules\Tests\PHPStan\Rule\SingleArgEventDispatchRule\Fixture;
 
-use TomasVotruba\Handyman\Tests\PHPStan\Rule\SingleArgEventDispatchRule\Source\NotEventDispatcher;
+use Symplify\PHPStanRules\Tests\PHPStan\Rule\SingleArgEventDispatchRule\Source\NotEventDispatcher;
 
 final class SkipUnrelatedDispatch
 {
