+++ tests/Rules/Symfony/SingleArgEventDispatchRule/Source/NotEventDispatcher.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace TomasVotruba\Handyman\Tests\PHPStan\Rule\SingleArgEventDispatchRule\Source;
+namespace Symplify\PHPStanRules\Tests\PHPStan\Rule\SingleArgEventDispatchRule\Source;
 
 final class NotEventDispatcher
 {
